12.6.1-9 baseline
Former-commit-id: 56ce907ff214df2b6aa51305b940d17f36d2a2fa [formerly 32fdf6ed3b3a2f96b37aaf91e5ba536de28a535a] [formerly 7c064b686806b5fb395b16e7c211d0bc066a6d84] [formerly 7c064b686806b5fb395b16e7c211d0bc066a6d84 [formerly 2c0ff851cab2365e537f56ca3b74a4a4ea53ead9]] [formerly 56ce907ff214df2b6aa51305b940d17f36d2a2fa [formerly 32fdf6ed3b3a2f96b37aaf91e5ba536de28a535a] [formerly 7c064b686806b5fb395b16e7c211d0bc066a6d84] [formerly 7c064b686806b5fb395b16e7c211d0bc066a6d84 [formerly 2c0ff851cab2365e537f56ca3b74a4a4ea53ead9]] [formerly 27afa6034501403e42fdd888959ddfa882728195 [formerly 7c064b686806b5fb395b16e7c211d0bc066a6d84 [formerly 2c0ff851cab2365e537f56ca3b74a4a4ea53ead9] [formerly 27afa6034501403e42fdd888959ddfa882728195 [formerly 8f93efb93ec2d244fe599c16b3af5b33784c0232]]]]]
Former-commit-id: 27afa6034501403e42fdd888959ddfa882728195
Former-commit-id: 9fc0cbac36f0797ce018215ab602f335125224da [formerly e528bb0df2c07391eb792a5d1fbe5e081cc0159e] [formerly dc50bc8443c4e16df4a51d61bdfd7abe7b264de2] [formerly 202bf137851885adaaf1ac7f8ae4b23be4084de6 [formerly 39c28d347eb5aeecbc48481fe6871268a89c0c2d] [formerly dc50bc8443c4e16df4a51d61bdfd7abe7b264de2 [formerly 4f7bb90bb0661624fcb4890b96ae85cd59d35d1c]]]
Former-commit-id: 94f256444a3b2e57847ac115fc0d1511e30f8e94 [formerly 38b944b7b39fc0e1e6c5f5e0171cfa20b64068e4] [formerly 37f2b6e1381bf16ad07ddcc4934316dabd46a8c2 [formerly d5fa61041bcc936c428601c247c6ee01fe143bec]]
Former-commit-id: 37f2b6e1381bf16ad07ddcc4934316dabd46a8c2
Former-commit-id: c74431ba29a6a47126786303d45a4f6b5bb97a9b

+++ cave/com.raytheon.uf.viz.d2d.nsharp/src/com/raytheon/uf/viz/d2d/nsharp/rsc/D2DNSharpResourceData.java
@@ -178,8 +178,8 @@ public abstract class D2DNSharpResourceData extends
             stnInfo.setRangestarttime(fcstTime);
         }
         if (coordinate != null) {
-            stnInfo.setLongitude((float) coordinate.x);
-            stnInfo.setLatitude((float) coordinate.y);
+            stnInfo.setLongitude(coordinate.x);
+            stnInfo.setLatitude(coordinate.y);
         }
         if (pointName != null) {
             stnInfo.setStnDisplayInfo(pointName + " "
